:zap: add http status code

+++ apiRoutes/team.js
@@ -11,13 +11,14 @@ router.get("/", async (req, res) => {
     let data = await team.findAll({
         include:[{
             model: user,
-            through: { attributes:[] }
+            through:{attributes:[]},
+            attributes:{exclude:["password"]}
         }]
     }).catch(err=>{
         console.log(err)
-        res.sendStatus(400)
+        return res.status(500).send("Data corrupted.")
     })
-    res.send(data)
+    res.status(200).send(data)
 })
 
 router.post("/add", async (req, res) => {
@@ -27,55 +28,73 @@ router.post("/add", async (req, res) => {
         name: req.body.name
     }).catch(err=>{
         console.log(err)
-        res.sendStatus(400)
+        return res.sendStatus(400)
     })
-    data.addUser(req.body.leader_ID)
-    res.send(data)
+    data.addAllTeam(req.body.leader_ID)
+    res.status(201).send(data)
 })
 
 router.patch("/update/:id", async (req,res)=>{
     let teamID = req.params.id
+    //Check if id is Integer or not
+    if(isNaN(teamID)) return res.status(400).send("ID should be number.")
+    //Check if id is found or not
+    let isFound = await team.findOne({where:{team_ID: teamID}}).catch(err=>{
+        console.log(err)
+        return res.sendStatus(500)
+    })
+    if (isFound == null) return res.status(404).send("Team not found.")
+    //Update
     let data = await team.update(req.body, {where:{team_ID: teamID}}).catch(err=>{
         console.log(err)
-        res.sendStatus(400)
+        return res.sendStatus(500)
     })
-    res.send(data)
+    res.status(200).send(data)
 })
 
 router.patch("/user/add/:id", async (req, res) => {
     let teamID = req.params.id
     let userlist = req.body.userlist
-
+    //Check if id is Integer or not
+    if(isNaN(teamID)) return res.status(400).send("ID should be number.")
     //Get team data from teamID
     let teamData = await team.findOne({where:{ team_ID: teamID }}).catch(err => {
         console.log(err)
-        res.sendStatus(404).send("Team not found.")
+        res.sendStatus(500)
     })
-
-    teamData.addUsers(userlist)
-    res.sendStatus(200)
+    //Check if id is found or not
+    if(teamData == null) return res.status(404).send("Team not found.")
+    teamData.addAllTeam(userlist)
+    res.status(200).send(teamData)
 })
 
 router.patch("/user/delete/:id", async (req, res) => {
     let teamID = req.params.id
     let userlist = req.body.userlist
-
+    //Check if id is Integer or not
+    if(isNaN(teamID)) return res.status(400).send("ID should be number.")
     //Get team data from teamID
     let teamData = await team.findOne({where:{ team_ID: teamID }}).catch(err => {
         console.log(err)
-        res.sendStatus(404).send("Team not found.")
+        res.sendStatus(500)
     })
-
-    teamData.removeUsers(userlist)
-    res.sendStatus(200)
+    //Check if id is found or not
+    if(teamData == null) return res.status(404).send("Team not found.")
+    teamData.removeAllTeam(userlist)
+    res.status(200).send(teamData)
 })
 
 router.delete("/delete/:id", async (req, res)=>{
     let teamID = req.params.id
+    //Check if id is Integer or not
+    if(isNaN(teamID)) return res.status(400).send("ID should be number.")
+    //Get team data from teamID and delete
     let data =  await team.findOne({where: {team_ID: teamID}}).then((result) => {
         return team.destroy({where:{team_ID: teamID}}).then(() => {return result})
     })
-    res.send(data)
+    //Check if id is found or not
+    if(data == null) return res.status(404).send("Team not found.")
+    res.status(200).send(data)
 })
 
 module.exports = router
